Rename files to mark them as browser or electron
