+++ extensions/typescript-language-features/src/extension.browser.ts
@@ -11,7 +11,7 @@ import { createLazyClientHost, lazilyActivateClient } from './lazyClientHost';
 import { noopRequestCancellerFactory } from './tsServer/cancellation';
 import { noopLogDirectoryProvider } from './tsServer/logDirectoryProvider';
 import { ITypeScriptVersionProvider, TypeScriptVersion, TypeScriptVersionSource } from './tsServer/versionProvider';
-import { WorkerServerProcess } from './tsServer/workerServerProcess';
+import { WorkerServerProcess } from './tsServer/serverProcess.browser';
 import API from './utils/api';
 import { CommandManager } from './utils/commandManager';
 import { TypeScriptServiceConfiguration } from './utils/configuration';
