+++ extensions/typescript-language-features/src/tsServer/serverProcess.electron.ts
